refactor : 전체적인 구조 refactor
