+++ src/main/java/kr/codesquad/Main.java
@@ -1,12 +1,10 @@
 package kr.codesquad;
 
-import kr.codesquad.LottoService.Lotto;
-
-import java.util.Scanner;
+import kr.codesquad.controller.LottoGameController;
 
 public class Main {
     public static void main(String[] args) {
-        App app = new App();
-        app.run();
+        LottoGameController lottoGameController = new LottoGameController();
+        lottoGameController.run();
     }
 }
