+++ src/main/java/kr/codesquad/domain/Lotto.java
@@ -1,4 +1,4 @@
-package kr.codesquad.LottoService;
+package kr.codesquad.domain;
 
 import java.util.*;
 import java.util.stream.Collectors;
@@ -9,7 +9,7 @@ public class Lotto {
 
     private Lotto(Set<LottoNumber> lotto) {
         if (lotto.size() != LOTTO_SIZE) {
-            throw new IllegalArgumentException("로또에는 6개의 숫자가 필요합니다.");
+            throw new IllegalArgumentException("로또에는 서로 다른 6개의 숫자가 필요합니다.");
         }
 
         this.lotto = lotto;
