+++ src/main/java/kr/codesquad/domain/LottoFactory.java
@@ -1,6 +1,6 @@
-package kr.codesquad.LottoService;
+package kr.codesquad.domain;
 
-import kr.codesquad.View.IOManager;
+import kr.codesquad.view.IOManager;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -9,7 +9,6 @@ import java.util.List;
 public class LottoFactory {
     private static final int LOTTO_START_NUMBER = 1;
     private static final int LOTTO_LAST_NUMBER = 45;
-    private static final String ILLEGAL_ARGUMENT_MESSAGE = "유효하지 않은 로또 종류입니다";
     private final List<Integer> numbers;
     private final IOManager ioManager;
 
@@ -32,6 +31,6 @@ public class LottoFactory {
         if (type.equals("winning")){
             return Lotto.ofComma(ioManager.inputWinningNumber());
         }
-        throw new IllegalArgumentException(ILLEGAL_ARGUMENT_MESSAGE);
+        throw new IllegalArgumentException("유효하지 않은 로또 종류입니다");
     }
 }
